Use Go primitives in internal engine interfaces

+++ src/legacy/briefing_window.go
@@ -53,7 +53,7 @@ func briefingCreateWindow() *gui.Window {
 	*memmap.PtrUint32(0x5D4594, 831284) = uint32(ox)
 	*memmap.PtrUint32(0x5D4594, 831288) = uint32(oy)
 	parent := GetClient().Cli().GUI.NewWindowRaw(nil, 56, 0, 0, int(nox_win_width), int(nox_win_height), briefingEvent(briefingBackgroundEvent))
-	legacyGlobals.dword_5d4594_831236 = C.uint32_t(uintptr(parent.C()))
+	legacyGlobals.dword_5d4594_831236 = C.uint32_t(uint32(uintptr(parent.C())))
 	if parent == nil {
 		return nil
 	}
