+++ src/legacy/client_ui_inventory_stack.go
@@ -56,11 +56,11 @@ func uiInventorySetAlternate(cell *uiInventoryCell) int {
 	dword_5d4594_1062480 = uint32(uiInventoryPointer(unsafe.Pointer(cell)))
 	uiInventoryClearAlternateFlags()
 	if cell == nil {
-		return int(nox_xxx_clientReportSecondaryWeapon_4BF010(0))
+		return int(nox_xxx_clientReportSecondaryWeapon_4BF010(int32(0)))
 	}
 	cell.Drawable.NetCode32 = cell.Codes[0]
 	cell.Alternate = 1
-	return int(nox_xxx_clientReportSecondaryWeapon_4BF010(C.int(uiInventoryPointer(cell.Drawable.C()))))
+	return int(nox_xxx_clientReportSecondaryWeapon_4BF010(int32(uiInventoryPointer(cell.Drawable.C()))))
 }
 
 //export nox_xxx_clientSetAltWeapon_461550
