+++ src/legacy/player_controls_movement.go
@@ -88,7 +88,7 @@ func controlDropBall(u *server.Object) int32 {
 		*controlPtr(it.CObj(), 520) = nil
 		GetServer().S().ObjSetOwner(nil, it)
 		nox_xxx_aud_501960(926, (*C.nox_object_t)(u.CObj()), 0, 0)
-		sub_4E8290(1, 0)
+		sub_4E8290(int8(1), int16(0))
 		return 1
 	}
 	return 0
@@ -149,7 +149,7 @@ func controlInputAttack(u *server.Object) {
 		if *controlByte(data, 108) != 0 || *controlByte(data, 109) == 0 {
 			setAttack()
 			nox_xxx_playerSetState_4FA020((*C.nox_object_t)(u.CObj()), 1)
-			nox_xxx_useByNetCode_53F8E0(inventoryInt(u), inventoryInt(it))
+			nox_xxx_useByNetCode_53F8E0(C.int(inventoryInt(u)), C.int(inventoryInt(it)))
 		} else if controlSubStamina(u, 45) != 0 {
 			*equipmentWord(data, 96) |= 2
 			setAttack()
